Parameters

General

JutulDarcy.FluidVolume — Type

FluidVolume()

Variable typically taken to be a parameter. Represents the per-cell volume that where multiphase flow can occur. For a well, this is the volume inside the well-bore free flow can occur. For a porous medium, this is the void space inside the pores that is, to some extent, connected and open to flow (effective pore-volume).

Reservoir parameters

Transmissibility

JutulDarcy.Transmissibilities — Type

Transmissibilities()

Variable/parameter used to define the cell-to-cell transmissibilities when using a two-point flux approximation scheme.

JutulDarcy.reservoir_transmissibility — Function

reservoir_transmissibility(d::DataDomain)
reservoir_transmissibility(d::DataDomain; version = :xyz)

Special transmissibility function for reservoir simulation that handles additional complexity present in industry grade models such as fault multipliers, net-to-gross etc. The input should be a DataDomain instance returned from reservoir_domain

The keyword argument version can be :xyz for permeability tensors that are aligned with coordinate directions or :ijk to interpreted the permeability as a diagonal tensor aligned with the logical grid directions. The latter choice is only meaningful for a diagonal tensor.

Other

JutulDarcy.TwoPointGravityDifference — Type

TwoPointGravityDifference()

Parameter representing the difference in gravity on an instance of Faces between two Cells. If the phase flux is written as

$v = - K \nabla (p + \rho g \nabla z)$

this term represent the discretized analogue of $\rho g \nabla z$.

Missing docstring.

Missing docstring for JutulDarcy.ConnateWater. Check Documenter's build log for details.

Missing docstring.

Missing docstring for JutulDarcy.EndPointScalingCoefficients. Check Documenter's build log for details.

Well parameters

Missing docstring.

Missing docstring for JutulDarcy.WellIndices. Check Documenter's build log for details.

Missing docstring.

Missing docstring for JutulDarcy.PerforationGravityDifference. Check Documenter's build log for details.