Secondary variables (properties)

Fluid systems

General

Relative permeabilities

JutulDarcy.BrooksCoreyRelativePermeabilities — Type

BrooksCoreyRelativePermeabilities(
    sys_or_nph::Union{MultiPhaseSystem, Integer},
    exponents = 1.0,
    residuals = 0.0,
    endpoints = 1.0
)

Secondary variable that implements the family of Brooks-Corey relative permeability functions. This is a simple analytical expression for relative permeabilities that has a limited number of parameters:

$K(S) = K_{max} * ((S - S_r)/(1 - S_r^{tot}))^N$

Fields

exponents: Exponents for each phase

residuals: Residual saturations for each phase

endpoints: Maximum relative permeability for each phase

residual_total: Total residual saturation over all phases

JutulDarcy.RelativePermeabilities — Type

RelativePermeabilities((kr1, kr2, ...))

A simple relative permeability implementation. Assumes that each phase has a relative permeability on the form:

$K_{r,phase} = F(S_{phase})$

Supports multiple fluid regions through the regions keyword.

Examples

Single region:

kr1 = S -> S^2
kr2 = S -> S^3

kr = RelativePermeabilities((kr1, kr2))

Two regions:

kr1_reg1 = S -> S^2
kr2_reg1 = S -> S^3

kr1_reg2 = S -> S^3
kr2_reg2 = S -> S^4

regions # should be a vector with one entry that is 1 or 2 for each cell in the domain

kr = RelativePermeabilities(((kr1_reg1, kr2_reg1), (kr1_reg2, kr2_reg2)), regions = regions)

JutulDarcy.ReservoirRelativePermeabilities — Type

ReservoirRelativePermeabilities(
    w = nothing, g = nothing, ow = nothing, og = nothing,
    scaling = NoKrScale, regions = nothing)

Relative permeability with advanced features for reservoir simulation. Includes features like rel. perm. endpoint scaling, connate water adjustment and separate phase pair relative permeabilites for the oil phase.

Fields

krw

krow

krog

krg

regions

phases

JutulDarcy.PhaseRelativePermeability — Type

PhaseRelativePermeability(s, k; label = :w, connate = s[1], epsilon = 1e-16)

Type that stores a sorted phase relative permeability table (given as vectors of equal length s and k):

$K_r = K(S)$

Optionally, a label for the phase, the connate saturation and a small epsilon value used to avoid extrapolation can be specified.

Missing docstring.

Missing docstring for EndPointScalingCoefficients. Check Documenter's build log for details.

Phase viscosities

JutulDarcy.DeckPhaseViscosities — Type

DeckPhaseViscosities(pvt, regions = nothing)

Secondary variable used to evaluate viscosities when a case is generated from a input file. Typically not instantiated in user scripts.

JutulDarcy.PhaseMassDensities — Type

Abstract type representing the evaluation of mass density of each phase (i.e. units of mass per units of volume, for each cell in the model domain.)

Immiscible flow

Phase densities

Shrinkage factors

JutulDarcy.DeckShrinkageFactors — Type

DeckShrinkageFactors(pvt, regions = nothing)

Secondary variable used to evaluate shrinkage factors when a case is generated from a input file. Typically not instantiated in user scripts.

JutulDarcy.ConstantCompressibilityDensities — Type

ConstantCompressibilityDensities(
    sys_or_nph::Union{MultiPhaseSystem, Integer},
    reference_pressure = 1.0,
    reference_density = 0.0,
    compressibility = 1.0
)

Secondary variable that implements a constant compressibility relationship for density. Given the reference pressure, compressibility and density at the reference pressure, each phase density can be computed as:

$ρ(S) = ρ_{ref} e^{(p - p_{ref})c}$

The constructor can take in either one value per phase or a single value for all phases for the reference pressure, compressibility and density at reference conditions.

Fields

reference_pressure: Reference pressure for each phase (where the reference densities are given)

reference_densities: Densities at the reference point

compressibility: Compressibility factor used when expanding around reference pressure, typically between 1e-3 and 1e-10

Black-oil flow

Compositional flow

JutulDarcy.PhaseMassFractions — Type

PhaseMassFractions(:liquid)

Variable that defines the component mass fractions in a specific phase.

Wells

JutulDarcy.TotalMass — Type

TotalMasses()

Variable that defines total mass of all components in each cell of the domain.