Jutul.simulator_config — Function

simulator_config(sim; info_level = 3, linear_solver = GenericKrylov())

Set up a simulator configuration object that can be passed onto simulate!.

There are many options available to configure a given simulator. The best way to get an overview of these possible configuration options is to instatiate the config without any arguments and inspecting the resulting table by calling simulator_config(sim) in the REPL.

Jutul.simulate! — Function

simulate!(sim::JutulSimulator, timesteps::AbstractVector; forces = nothing,
                                                               config = nothing,
                                                               initialize = true,
                                                               restart = nothing,
                                                               state0 = nothing,
                                                               parameters = nothing,
                                                               kwarg...)

Non-allocating (or perhaps less allocating) version of simulate!.

Arguments

initialize=true: Perform internal updates as if this is the first time

See also simulate for additional supported input arguments.

Jutul.simulate — Function

simulate(state0, model, timesteps, parameters = setup_parameters(model))
simulate(state0, model, timesteps, info_level = 3)
simulate(state0, model, timesteps; <keyword arguments>)

Simulate a set of timesteps with model for the given initial state0 and optionally specific parameters. Additional keyword arguments are passed onto simulator_config and simulate!. This interface is primarily for convenience, as all storage for the simulator is allocated upon use and discared upon return. If you want to perform multiple simulations with the same model it is advised to instead instantiate Simulator and combine it with simulate!.

Arguments

state0::Dict: initial state, typically created using setup_state for the model in use.

model::JutulModel: model that describes the discretized system to solve, for example SimulationModel or MultiModel.

timesteps::AbstractVector: Vector of desired report steps. The simulator will perform time integration until sum(timesteps) is reached, providing outputs at the end of each report step.

parameters=setup_parameters(model): Optional overrides the default parameters for the model.

forces=nothing: Either nothing (for no forces), a single set of forces from setup_forces(model) or a Vector of such forces with equal length to timesteps.

restart=nothing: If an integer is provided, the simulation will attempt to restart from that step. Requires that output_path is provided here or in the config.

config=simulator_config(model): Configuration Dict that holds many fine grained settings for output, linear solver, time-steps, outputs etc.

Additional arguments are passed onto simulator_config.

See also simulate!, Simulator, SimulationModel, simulator_config.

simulate(state0, sim::JutulSimulator, timesteps::AbstractVector; parameters = nothing, kwarg...)

Simulate a set of timesteps with simulator for the given initial state0 and optionally specific parameters.

Jutul.JutulSystem — Type

Abstract type for the physical system to be solved.

Jutul.JutulMesh — Type

A mesh is a type of domain that has been discretized. Abstract subtype.

Jutul.JutulDomain — Type

Abstract type for domains where equations can be defined

Jutul.VectorVariables — Type

Abstract type for vector variables (more than one entry per entity, for example saturations or displacements)

Jutul.MultiModel — Type

MultiModel(models)

A model variant that is made up of many named submodels, each a fully realized SimulationModel.

models should be a NamedTuple or Dict{Symbol, JutulModel}.

Jutul.SimulationModel — Type

SimulationModel(domain, system; <kwarg>)

Instantiate a model for a given system discretized on the domain.

SimulationModel(g::JutulMesh, system; discretization = nothing, kwarg...)

Type that defines a simulation model - everything needed to solve the discrete equations.

The minimal setup requires a JutulMesh that defines topology together with a JutulSystem that imposes physical laws.

Jutul.Simulator — Type

Simulator(model; <kwarg>)

Set up a simulator object for a model that can be used by simulate!. To avoid manually instantiating the simulator, the non-mutating simulate interface can be used instead.

Jutul.setup_parameters — Function

setup_parameters(model::JutulModel; name = value)

Set up a parameter storage for a given model with values for the parameter defined in the model.

Arguments

name=value: The name of the parameter together with the value(s) of the parameter.

A scalar (or short vector of the right size for VectorVariables) will be repeated over the entire domain, while a vector (or matrix for VectorVariables) with length (number of columns for VectorVariables) equal to the entity count (for example, number of cells for a cell variable) will be used directly.

Jutul.setup_state — Function

setup_state(model::JutulModel, name1 = value1, name2 = value2)

Set up a state for a given model with values for the primary variables defined in the model. Normally all primary variables must be initialized in this way.

Arguments

name=value: The name of the primary variable together with the value(s) used to initialize the primary variable.

A scalar (or short vector of the right size for VectorVariables) will be repeated over the entire domain, while a vector (or matrix for VectorVariables) with length (number of columns for VectorVariables) equal to the entity count (for example, number of cells for a cell variable) will be used directly.

Note: You likely want to overload [setup_state!]@ref for a custom model instead of setup_state

Jutul.setup_forces — Function

setup_forces(model::JutulModel; force_name = force_value)

Set up forces for a given model. Keyword arguments varies depending on what the model supports.

JutulDarcy.MIN_ACTIVE_WELL_RATE — Constant

(Absolute) Minimum well rate for a well that is not disabled.

JutulDarcy.MIN_INITIAL_WELL_RATE — Constant

(Absolute) Minimum initial rate for wells when controls are updated.

JutulDarcy.BlackOilX — Method

BlackOilX(sys::BlackOilVariableSwitchingSystem, p; sw = 0.0, so = 0.0, sg = 0.0, rs = 0.0, rv = 0.0)

High level initializer for the black oil unknown degree of freedom. Will try to fill in the gaps unless system is really underspecified.

JutulDarcy.CPRPreconditioner — Type

CPRPreconditioner(p = default_psolve(), s = ILUZeroPreconditioner(); strategy = :quasi_impes, weight_scaling = :unit, update_frequency = 1, update_interval = :iteration, partial_update = true)

Construct a constrained pressure residual (CPR) preconditioner.

By default, this is a AMG-BILU(0) version (algebraic multigrid for pressure, block-ILU(0) for the global system).

JutulDarcy.CPRPreconditioner — Type

Constrained pressure residual

JutulDarcy.DisabledControl — Type

DisabledControl()

Control that disables a well. If a well is disabled, it is disconnected from the surface network and no flow occurs between the well and the top side. Mass transfer can still occur inside the well, and between the well and the reservoir.

See also ProducerControl, InjectorControl.

JutulDarcy.DisabledTarget — Type

DisabledTarget(q)

Disabled target used when a well is under DisabledControl() only.

JutulDarcy.FlowBoundaryCondition — Type

FlowBoundaryCondition(
cell,
pressure = DEFAULT_MINIMUM_PRESSURE, 
temperature = 298.15;
fractional_flow = nothing,
density = nothing,
trans_flow = 1e-12,
trans_thermal = 1e-6
)

Boundary condition for constant values (pressure/temperature)

JutulDarcy.HistoricalReservoirVoidageTarget — Type

HistoricalReservoirVoidageTarget(q, weights)

Historical RESV target for history matching cases.

JutulDarcy.ImmiscibleSaturation — Type

A single saturation that represents the "other" phase in a three phase compositional system where two phases are predicted by an EoS

JutulDarcy.ImmiscibleSystem — Method

ImmiscibleSystem(phases; reference_densities = ones(length(phases)))
ImmiscibleSystem((LiquidPhase(), VaporPhase()), (1000.0, 700.0))

Set up an immiscible system for the given phases with optional reference densitites. This system is easy to specify with Pressure and Saturations as the default primary variables. Immiscible system assume that there is no mass transfer between phases and that a phase is uniform in composition.

JutulDarcy.InjectorControl — Type

InjectorControl(target, mix, [density])

Well control that specifies injection into the reservoir. target specifies the type of target and mix defines the injection mass fractions for all species in the model during injection.

For example, for a three-component system made up of CO2, H2O and H2, setting [0.1, 0.6, 0.3] would mean that the injection stream would contain 1 part CO2, 6 parts H2O and 3 parts H2 by mass. For an immiscible system (e.g. LiquidPhase(), VaporPhase()) the species corresponds to phases and [0.3, 0.7] would mean a 3 to 7 mixture of liquid and vapor by mass.

The density of the injected fluid at surface conditions is given by density which is defaulted to 1.0 if not given.

See also ProducerControl, DisabledControl.

JutulDarcy.MinimalTPFAGrid — Type

MinimalTPFAGrid(ϕ, N)

Generate a minimal grid suitable only for two-point flux discretization (TPFA) for given pore-volumes ϕ and a neighborship matrix N with size (2, n) where n is the number of internal faces.

JutulDarcy.MixedWellSegmentFlow — Type

Two point approximation with flux for wells

JutulDarcy.MultiPhaseCompositionalSystemLV — Type

MultiPhaseCompositionalSystemLV(equation_of_state, phases = (LiquidPhase(), VaporPhase()); reference_densities = ones(length(phases)), other_name = "Water")

Set up a compositional system for a given equation_of_state from MultiComponentFlash.

JutulDarcy.MultiSegmentWell — Method

MultiSegmentWell(reservoir_cells, volumes, centers;
                N = nothing,
                name = :Well,
                perforation_cells = nothing,
                segment_models = nothing,
                reference_depth = 0,
                dz = nothing,
                surface_conditions = default_surface_cond(),
                accumulator_volume = mean(volumes),
                )

Create well perforated in a vector of reservoir_cells with corresponding volumes and cell centers.

NOTE: setup_vertical_well or setup_well are the recommended way of setting up wells.

JutulDarcy.OverallMoleFractions — Method

OverallMoleFractions(;dz_max = 0.2)

Overall mole fractions definition for compositional. dz_max is the maximum allowable change in any composition during a single Newton iteration.

JutulDarcy.Perforations — Type

Perforations are connections from well cells to reservoir vcells

JutulDarcy.PhaseMassDensities — Type

Mass density of each phase

JutulDarcy.Pressure — Method

Pressure(; max_abs = nothing, max_rel = 0.2, scale = 1e8, maximum = Inf, minimum = DEFAULT_MINIMUM_PRESSURE)

Pressure variable definition. max_abs/max_rel maximum allowable absolute/relative change over a Newton iteration, scale is a "typical" value used to regularize the linear system, maximum the largest possible value and minimum the smallest.

JutulDarcy.ProducerControl — Type

ProducerControl(target)

Well control for production out of the reservoir. target specifies the type of target (for example BottomHolePressureTarget()).

See also DisabledControl, InjectorControl.

JutulDarcy.ReservoirRelativePermeability — Type

Interpolated multiphase rel. perm. that is simple (single region, no magic for more than two phases)

JutulDarcy.ReservoirVoidageTarget — Type

ReservoirVoidageTarget(q, weights)

RESV targets with weights for each pseudo-component

JutulDarcy.Saturations — Method

Saturations(;ds_max = 0.2)

Saturations as primary variable. ds_max controls maximum allowable saturation change between two Newton iterations.

JutulDarcy.SegmentWellBoreFrictionHB — Type

Hagedorn and Brown well bore friction model for a segment.

JutulDarcy.SimpleWell — Method

SimpleWell(reservoir_cells)

Set up a simple well.

NOTE: setup_vertical_well or setup_well are the recommended way of setting up wells.

JutulDarcy.SinglePhaseRateTarget — Type

SinglePhaseRateTarget(q, phase)

Single-phase well target with value q specified for phase.

Examples

julia> SinglePhaseRateTarget(0.001, LiquidPhase())
SinglePhaseRateTarget of 0.001 [m^3/s] for LiquidPhase()

JutulDarcy.SinglePhaseSystem — Type

SinglePhaseSystem(phase = LiquidPhase(); reference_density = 1.0)

A single-phase system that only solves for pressure.

JutulDarcy.SourceTerm — Method

SourceTerm(cell, value; fractional_flow = [1.0], type = MassSource)

Create source term in given cell with given total value. The optional fractional_flow argument controls how this term is divided over components if used for inflow.

JutulDarcy.StandardBlackOilSystem — Method

StandardBlackOilSystem(; rs_max = nothing,
                         rv_max = nothing,
                         phases = (AqueousPhase(), LiquidPhase(), VaporPhase()),
                         reference_densities = [786.507, 1037.84, 0.969758])

Set up a standard black-oil system. Keyword arguments rs_max and rv_max can either be nothing or callable objects / functions for the maximum Rs and Rv as a function of pressure. phases can be specified together with reference_densities for each phase.

NOTE: For the black-oil model, the reference densities significantly impact many aspects of the PVT behavior. These should generally be set consistently with the other properties.

JutulDarcy.SurfaceGasRateTarget — Type

SurfaceGasRateTarget(q)

Well target of specified gas rate with value q at surface conditions.

JutulDarcy.SurfaceLiquidRateTarget — Type

SurfaceLiquidRateTarget(q)

Well target of specified liquid rate with value q (liquid/oil and water, but not gas) at surface conditions.

JutulDarcy.SurfaceOilRateTarget — Type

SurfaceOilRateTarget(q)

Well target of specified oil rate with value q at surface conditions.

JutulDarcy.SurfaceWaterRateTarget — Type

SurfaceWaterRateTarget(q)

Well target of specified water rate with value q at surface conditions.

JutulDarcy.TabulatedRelPermSimple — Type

Interpolated multiphase rel. perm. that is simple (single region, no magic for more than two phases)

JutulDarcy.TotalRateTarget — Type

TotalRateTarget(q)

Well target of specified total rate of all phases with value q at surface conditions.

JutulDarcy.component_mass_fluxes! — Function

component_mass_fluxes!(q, face, state, model, kgrad, upw)

Implementation of component fluxes for a given system for a given face.

JutulDarcy.full_well_outputs — Method

full_well_outputs(model, states, forces; targets = available_well_targets(model.models.Reservoir), shortname = false)

Get the full set of well outputs after a simulation has occured, for plotting or other post-processing.

JutulDarcy.generate_phase_indices — Function

generate_phase_indices(phases, canonical = (a = AqueousPhase(), l = LiquidPhase(), v = VaporPhase()))

Generate mapping for canonical ordering of phases.

JutulDarcy.plot_reservoir_simulation_result — Method

plot_reservoir_simulation_result(model::MultiModel, res::ReservoirSimResult; wells = true, reservoir = true)

Plot a reservoir simulation result. If wells=true well curves will be shown interactively. If reservoir=true the reservoir quantities will be visualized in 3D. These options can be combined.

JutulDarcy.replace_target — Function

replace_target(ctrl, new_target)

Create new well control using ctrl as a template that operates under new_target.

JutulDarcy.reservoir_domain — Method

reservoir_domain(g; permeability = 9.869232667160130e-14, porosity = 0.1, kwarg...)

Set up a DataDomain instance for given mesh or other representation g. permeability and porosity are then added to the domain. If scalars are passed, they are expanded to cover all cells. Arrays are asserted to match all cells. Permeability is either one value per cell (diagonal scalar), one value per dimension given in each row (for a diagonal tensor) or a vector that represents a compact full tensor representation (6 elements in 3D, 3 in 2D).

JutulDarcy.reservoir_linsolve — Function

reservoir_linsolve(model, precond = :cpr; <keyword arguments>)

Set up iterative linear solver for a reservoir model from setup_reservoir_model.

Arguments

model: Reservoir model that will linearize the equations for the linear solver

precond=:cpr: Preconditioner type to use: Either :cpr (Constrained-Pressure-Residual) or :ilu0 (block-incomplete-LU) (no effect if solver = :direct).

v=0: verbosity (can lead to a large amount of output)

solver=:bicgstab: the symbol of a Krylov.jl solver (typically :gmres or :bicgstab)

update_interval=:once: how often the CPR AMG hierarchy is reconstructed (:once, :iteration, :ministep, :step)

update_interval_partial=:iteration: how often the pressure system is updated in CPR

max_coarse: max size of coarse level if using AMG

cpr_type=nothing: type of CPR (:true_impes, :quasi_impes or nothing for automatic)

partial_update=true: perform partial update of CPR preconditioner outside of AMG update (see above)

rtol=1e-3: relative tolerance for the linear solver

max_iterations=100: limit for linear solver iterations

Additional keywords are passed onto the linear solver constructor.

JutulDarcy.setup_reservoir_forces — Method

setup_reservoir_forces(model; control = nothing, limits = nothing, set_default_limits = true, <keyword arguments>)

Set up driving forces for a reservoir model with wells

JutulDarcy.setup_reservoir_model — Method

setup_reservoir_model(reservoir, system; wells = [], <keyword arguments>)
setup_reservoir_model(reservoir, system; wells = [], context = DefaultContext(), reservoir_context = nothing, backend = :csc, <keyword arguments>)

Set up a reservoir MultiModel for a given reservoir DataDomain typically set up from reservoir_domain and an optional vector of wells that are created using setup_vertical_well and setup_well.

The routine automatically sets up a facility and couples the wells with the reservoir and that facility.

JutulDarcy.setup_reservoir_simulator — Function

setup_reservoir_simulator(models, initializer, parameters = nothing; <keyword arguments>)

Arguments

models: either a single model or a Dict with the key :Reservoir for multimodels

initializer: used to setup state0, must be compatible with model

parameters: initialized parameters, must be compatible with model if provided

linear_solver=:bicgstab: iterative solver to use (provided model supports it)

precond=:cpr: preconditioner for iterative solver: Either :cpr or :ilu0.

rtol=1e-3: relative tolerance for linear solver

initial_dt=3600*24.0: initial time-step in seconds (one day by default)

target_its=8: target number of nonlinear iterations per time step

offset_its=1: dampening parameter for time step selector where larger values lead to more pessimistic estimates.

tol_cnv=1e-3: maximum allowable point-wise error (volume-balance)

tol_mb=1e-7: maximum alllowable integrated error (mass-balance)

specialize=false: use deep specialization of storage for faster execution, but significantly more compile time

Additional keyword arguments are passed onto simulator_config.

JutulDarcy.setup_reservoir_state — Method

setup_reservoir_state(model, <keyword arguments>)
# Ex: For immiscible two-phase
setup_reservoir_state(model, Pressure = 1e5, Saturations = [0.2, 0.8])

Convenience constructor that initializes a state for a MultiModel set up using setup_reservoir_model. The main convenience over setup_state is only the reservoir initialization values need be provided: wells are automatically initialized from the connected reservoir cells.

JutulDarcy.setup_vertical_well — Method

setup_vertical_well(g, K, i, j; heel = 1, toe = grid_dims_ijk(g)[3], kwarg...)

Set up a vertical well for given grid g and permeability K at logical indices i, j perforating all cells starting at k-logical index heel to toe.

JutulDarcy.setup_vertical_well — Method

setup_vertical_well(D::DataDomain, i, j; <kwarg>)

Set up a vertical well with a DataDomain input that represents the porous medium / reservoir where the wells it to be placed.

JutulDarcy.setup_well — Method

setup_well(D::DataDomain, reservoir_cells; skin = 0.0, Kh = nothing, radius = 0.1, dir = :z)

Set up a well in reservoir_cells with given skin factor and radius. The order of cells matter as it is treated as a trajectory.

JutulDarcy.simulate_mrst_case — Method

simulate_mrst_case(file_name; kwarg...)

Simulate a MRST case from file_name as exported by writeJutulInput in MRST.

Arguments

file_name::String: The path to a .mat file that is to be simulated.

extra_outputs::Vector{Symbol} = [:Saturations]: Additional variables to output from the simulation.

write_output::Bool = true: Write output (in the default JLD2 format)

output_path = nothing: Directory for output files. Files will be written under this directory. Defaults to the folder of file_name.

write_mrst = true: Write MRST compatible output after completed simulation that can be read by readJutulOutput in MRST.

backend=:csc: choice of backend for linear systems. :csc for default Julia sparse, :csr for experimental parallel CSR.

verbose=true: print some extra information specific to this routine upon calling

nthreads=Threads.nthreads(): number of threads to use

linear_solver=:bicgstab: name of Krylov.jl solver to use, or :direct (for small cases only)

info_level=0: standard Jutul info_level. 0 for minimal printing, -1 for no printing, 1-5 for various levels of verbosity

Additional input arguments are passed onto setup_reservoir_simulator and simulator_config if applicable.

JutulDarcy.two_phase_compositional_mass! — Method

Update total masses for two-phase compositional

JutulDarcy.two_phase_compositional_mass! — Method

Update total masses for two-phase compositional where another immiscible phase is present

JutulDarcy.update_total_masses! — Function

update_total_masses!(totmass, tv, model, arg..., ix)

Update total masses for a given system. Number of input arguments varies based on physical system under consideration.

JutulDarcy.well_mismatch — Method

well_mismatch(qoi, wells, model_f, states_f, model_c, state_c, dt, step_no, forces; <keyword arguments>)

Compute well mismatch for a set of qoi's (well targets) and a set of well symbols.

JutulDarcy.well_output — Function

well_output(model, states, well_symbol, forces, target = BottomHolePressureTarget)

Get a specific well output from a valid operational target once a simulation is completed an states are available.

JutulDarcy.well_target — Method

Well target contribution from well itself (bhp)

JutulDarcy.well_target — Method

Well target contribution from well itself (disabled, zero value)

JutulDarcy.well_target — Method

Well target contribution from well itself (surface volume, injector)

JutulDarcy.well_target — Method

Well target contribution from well itself (surface volume, injector)

JutulDarcy.well_target — Method

Well target contribution from well itself (RESV, producer)

JutulDarcy.well_target — Method

Well target contribution from well itself (surface volume, producer)